working git add .
