+++ NICE.Identity.Management/ClientApp/src/containers/UsersPage.tsx
@@ -1,8 +1,9 @@
 import React, { Component } from "react";
+import { observer } from "mobx-react";
 import { GridContainer } from "./GridContainer";
-import { UserStore, UserProfileStore } from "../store";
+import { UserStore, UserProfileStore } from "../store/User";
 
-
+@observer
 export class UsersPage extends Component {
   userProfileStore = new UserProfileStore();
   userStore = new UserStore();
@@ -28,7 +29,7 @@ export class UsersPage extends Component {
   render() {
     const loggedUser = this.userProfileStore.userProfile;
     return loggedUser.roles &&
-      loggedUser.roles.find(item => item === "Administrator") ? (
+      loggedUser.roles.find((item: string) => item === "administrator") ? (
       <GridContainer store={this.userStore} columnDefs={this.columnDefs} />
     ) : (
       <div>no role</div>
