+++ NICE.Identity.Management/ClientApp/src/store/User/User.ts
@@ -1,11 +1,10 @@
-import { Store } from "../";
+import { Store } from "../CommonStore";
 import { User } from "../../models/User";
 import { url } from "../../constants/urls";
 
 export class UserStore extends Store<User> {
   constructor() {
-    super();
-    this.url = url.users
+    super(url.users);
   }
   // specific methods for the User store
 }
